pyproject.toml
[build-system]
requires = ["setuptools>=64", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "bayzflow"
version = "0.1.1"
description = "Bayesian deep learning framework that converts any PyTorch model into a calibrated Bayesian (Pyro) model with uncertainty quantification"
readme = "README.md"
authors = [
    { name = "BayzFlow Contributors" }
]
license = { text = "MIT" }
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Scientific/Engineering :: Medical Science Apps.",
]
requires-python = ">=3.8"
dependencies = [
    "torch>=1.13.0",
    "pyro-ppl>=1.8.0",
    "numpy>=1.21.0",
    "pyyaml>=5.4.0",
]

[project.optional-dependencies]
fx = [
    "yfinance>=0.2.0",
    "pandas>=1.3.0",
    "ta>=0.10.0",
]

medical = [
    "monai>=1.0.0",
    "nibabel>=3.2.0",
]

viz = [
    "matplotlib>=3.5.0",
]

dev = [
    "pytest>=7.0.0",
    "black>=22.0.0",
    "isort>=5.10.0",
    "flake8>=4.0.0",
]

all = [
    "yfinance>=0.2.0",
    "pandas>=1.3.0",
    "ta>=0.10.0",
    "monai>=1.0.0",
    "nibabel>=3.2.0",
    "matplotlib>=3.5.0",
    "pytest>=7.0.0",
    "black>=22.0.0",
    "isort>=5.10.0",
    "flake8>=4.0.0",
]

[project.urls]
Homepage = "https://github.com/yourusername/bayzflow"
Documentation = "https://github.com/yourusername/bayzflow"
Repository = "https://github.com/yourusername/bayzflow"
"Bug Tracker" = "https://github.com/yourusername/bayzflow/issues"

[tool.setuptools]
packages = [
  "bayzflow",
  "bayzflow.core",
  "bayzflow.models",
  "bayzflow.examples",
  "bayzflow.examples.fx",
  "bayzflow.examples.fx.fx_utils",
  "bayzflow.examples.monai_examples",
  "bayzflow.examples.CiFAR",
]

[tool.setuptools.package-data]
bayzflow = ["*.yaml"]

[tool.black]
line-length = 100
target-version = ['py38', 'py39', 'py310', 'py311']

[tool.isort]
profile = "black"
line_length = 100

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
